+++ src/prompts/schemas/remote-remove.js
@@ -5,12 +5,12 @@
  */
 export default {
   properties: {
-    shoudRemove: {
+    shouldRemove: {
       required: true,
       description: "are you sure you would like to proceed with this operation? (yes[y]/no[n])'",
       message: 'please use yes or no.',
       type: 'string',
-      conform(value) {
+      conform(value: string) {
         return (
           value.toLowerCase() === 'y' ||
           value.toLowerCase() === 'n' ||