+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased
 
+- `bit remove` removes components from a remote scope only when `--remote` flag is used
 - fix bit-install to work from an inner directory
 - improve external test and build errors to show the stack
 - improve output of import command to show the imported versions